refectored contract
