+++ contracts/engagement/src/core/mod.rs
@@ -1,0 +1,10 @@
+pub mod user;
+pub mod admin;
+pub mod escrow;
+pub mod dispute;
+pub mod milestone;
+
+pub use user::*;
+pub use escrow::*;
+pub use dispute::*;
+pub use milestone::*;
