+++ contracts/engagement/src/events/mod.rs
@@ -1,0 +1,6 @@
+pub mod handler;
+
+pub (crate) use handler::{
+    escrows_by_engagement_id,
+    balance_retrieved_event,
+};
