+++ contracts/engagement/src/storage/types.rs
